# be/services/docs/audit-service/swagger.yaml
basePath: /api/v1
info:
  contact: {}
  description: API documentation for audit-service.
  title: Base Golang Audit Service API
  version: "1.0"
paths:
...
  /rbac/permissions:
    get:
      parameters:
      - description: Include hidden permissions
        in: query
        name: include_hidden
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - RBAC
  /rbac/roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - RBAC
    post:
      consumes:
      - application/json
      parameters:
      - description: Create role payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - RBAC
  /rbac/roles/{roleId}:
    delete:
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - RBAC
    put:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: Update role payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - RBAC
  /rbac/roles/{roleId}/permissions:
    get:
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get role permissions
      tags:
      - RBAC
    put:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: Set role permissions payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set role permissions
      tags:
      - RBAC
  /rbac/users/{userId}/role:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Assign role payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign user role
      tags:
      - RBAC
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"